refactor(types): split src/types/index.ts by domain
Splits the 458-LOC monolithic types file into six domain modules:

  - api.ts          connection, settings, ROM lookup/metadata
  - sync.ts         platform/collection sync, preview/plan/apply
  - downloads.ts    per-ROM download queue + progress events
  - firmware.ts     firmware + BIOS + AvailableCore
  - saves.ts        save sync, conflicts, slots, setup wizard
  - achievements.ts RetroAchievements data shapes

index.ts becomes a thin re-export façade so every existing
'from "../types"' import keeps working unchanged. Consumers that
want narrow imports can deep-import the domain module
('from "../types/saves"').

No behavior change; no consumer code modified.

Closes #609
